+++ trackintel/io/postgis.py
@@ -10,6 +10,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.types import JSON
 
 import trackintel as ti
+from trackintel.io.util import _index_warning_default_none
 
 
 def _handle_con_string(func):
@@ -40,6 +41,7 @@ def _handle_con_string(func):
     return wrapper
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_positionfixes_postgis(
     sql,
@@ -140,6 +142,7 @@ def write_positionfixes_postgis(
     )
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_triplegs_postgis(
     sql,
@@ -239,6 +242,7 @@ def write_triplegs_postgis(
     )
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_staypoints_postgis(
     sql,
@@ -340,6 +344,7 @@ def write_staypoints_postgis(
     )
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_locations_postgis(
     sql,
@@ -461,6 +466,7 @@ def write_locations_postgis(
     )
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_trips_postgis(
     sql,
@@ -590,6 +596,7 @@ def write_trips_postgis(
         )
 
 
+@_index_warning_default_none
 @_handle_con_string
 def read_tours_postgis(
     sql,
